refactor: Extract emitRequest to utils
